Add patch-package and update rn-nodeify
Adding `patch-package` to manage local dependency code changes. This is used to modify Userbase's hash algorithm name, changing it to lowercase 'sha-256' to work with browserify-sign which is used by rn-nodeify. The change will be made automatically when installing dependencies, triggered by the `postinstall` script in `package.json`.
Also updating `rn-nodeify` to 10.3.0 which includes a PR from @Fubinator.